refactor : 이익률의 역할만 수행하는 ProfitRate 객체 추가
